Update Item Request

+++ app/Http/Controllers/ItemRequestController.php
@@ -15,6 +15,7 @@ use App\Models\ItemRequest;
 use App\Models\SupplierType;
 use App\Models\SettingCompany;
 use App\Models\ProductSupplier;
+use App\Models\PotentialVendor;
 use App\Models\SupplierCategory;
 
 use App\Jobs\SentMessageToVendor;
@@ -102,6 +103,11 @@ class ItemRequestController extends Controller
 
     public function edit(ItemRequest $itemRequest)
     {
+        if(!$itemRequest->action_permission) 
+        {
+            return redirect()->back()->with('error', 'Anda tidak memiliki akses untuk mengedit data ini.');
+        }
+
         $categories = SupplierCategory::byCompany(Auth::user()->company_id)->get();
 
         $settingCompany = SettingCompany::byCompany(Auth::user()->company_id)->where('menu','wablas')->get()->pluck('field_value','field_title');
@@ -138,6 +144,39 @@ class ItemRequestController extends Controller
             'qty' => 'required|numeric|min:1',
         ]);
         
+                // Kelola PotentialVendor berdasarkan product_supplier_id
+        if (!empty($request->product_supplier_id)) 
+        {
+            $inputVendorIds = $request->product_supplier_id ?? [];
+
+            // Step 1: Ambil ID existing dari relasi
+            $existingVendorIds = $itemRequest->potentialVendors->pluck('product_supplier_id')->toArray();
+
+            // Step 2: Simpan atau update data baru
+            foreach ($inputVendorIds as $vendorId) {
+                PotentialVendor::firstOrCreate([
+                    'company_id' => $itemRequest->company_id,
+                    'item_request_id' => $itemRequest->id,
+                    'product_supplier_id' => $vendorId,
+                ], [
+                    'responded' => false
+                ]);
+            }
+
+            // Step 3: Hapus data lama yang tidak ada dalam input baru
+            $vendorsToDelete = array_diff($existingVendorIds, $inputVendorIds);
+
+            if (!empty($vendorsToDelete)) {
+                PotentialVendor::where('item_request_id', $itemRequest->id)
+                    ->whereIn('product_supplier_id', $vendorsToDelete)
+                    ->delete();
+            }
+        } 
+        // else 
+        // {
+        //     $itemRequest->potentialVendors()->delete();
+        // }
+        
         $validated['supplier_type_id'] = $request->type;
 
         $settingCompany = SettingCompany::byCompany(Auth::user()->company_id)->where('menu','wablas')->get()->pluck('field_value','field_title');
@@ -159,7 +198,7 @@ class ItemRequestController extends Controller
         // dispatch(new ProcessItemRequestCreated($itemRequest->id));
         $itemRequest->update($validated);
 
-        return redirect()->route('item-request.index')->with('success', 'Request updated.');
+        return redirect()->route('item-request.show',$itemRequest->id)->with('success', 'Request updated.');
     }
 
     public function destroy(ItemRequest $itemRequest)
@@ -168,6 +207,12 @@ class ItemRequestController extends Controller
         {
             return redirect()->route('item-request.index')->with('error', 'Error: Request sudah selesai, tidak dapat dihapus.');
         }
+
+        if(!$itemRequest->action_permission) 
+        {
+            return redirect()->back()->with('error', 'Anda tidak memiliki akses untuk mengedit data ini.');
+        }
+        
         $itemRequest->delete();
         return redirect()->route('item-request.index')->with('success', 'Request deleted.');
     }
@@ -237,9 +282,15 @@ class ItemRequestController extends Controller
 
         foreach ($data->data as $index => $item) 
         {
+            $item->item_name = '<a href="'.route('item-request.show',$item->id).'">'.$item->item_name.'</a>';
             $item->estimated_price = 'Rp. '.number_format($item->estimated_price, 0,',','.'); // Format angka dengan 2 desimal
             $item->status = $item->status_badge;
+            if(!$item->action_permission)
+            {
+                $item->action = '<span class="badge badge-danger">CLOSED</span>';
+            }
         }
+
         
         return response()->json($data);
     }
